ci(KFLUXVNGD-389): drop reviewdog
Its going to take more effort to to make reciewdog work properly with
GitHub (E.g. it needs to run in a "pull_request_target" workflow and
avoid executing any of the code from the PR), so dropping it for now,
but keeping it in the Git history for when we want to try again.

Signed-off-by: Barak Korren <bkorren@redhat.com>