chore: multiple envs in helm values file
Adding a variable in the values file for setting the target env for
deployment. The task that builds the helm chart will then update
variables relevant for the deployment of the different environments,
and Helm helpers implemented as part of this change will select the
relevant values based on the deployment environment.

This will allow using a single values file for all environments,
constructing it during build time and releasing it as-is, rather than
having to modify it between build and release phases.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>