fix: openshift configuration for squid
The existing setup caused some issues when trying to install on
openshift, specifically how the initContainer was set up to work.

Instead, we move the database initialization to the entrypoint script
and avoid using the initContainer altogether.

This is to prevent the need in initContainer that also requires
some permissions that are not as trivial to provide on openshift.

Other than that, to allow only installing certificates without
installing cert-manager on openshift, we also add selfsigned-certificate
flag.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>