chore: add PR template
The underlying assumption is that much of the PRs that we submit is
generated by AI, but we need to somehow verify that changes make sense.

The idea is to foster a culture in which we only merge PRs for which we
understand what they do and how they do it, and that they only contain
the things that they need to contain, rather than everything that AI
chose to generate.

Keeping the checklist as short as possible is essential for people to
actually try and follow it rather than to ignore or to manipulate it.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>