Add store ID helper for quay CDN requests
This fixes caching for image layer blobs pulled from Quay backed registries

A rewrite-url helper cannot be used to pull images from registry.access.redhat.com
because the CDN urls contain temporary credentials with a signature calculated using
the host header. Squid cannot rewrite the host header for https connections and so
the rwwritten request fails.

To ensure a request is properly authorized to access an already cached blob, the
store id helper itself makes an initial request to ensure access. If access is granted
(a 200 response), the computed cache key is the original request URL sans query params.
Otherwise it falls back to the URL containing the temporary credentials which should
never result in a cache hit.

Signed-off-by: Alex Misstear <amisstea@redhat.com>